# dfras-docs/dfras-api-swagger.yaml
openapi: 3.0.3
info:
  title: DFRAS API
  description: |
    Delivery Failure Root Cause Analysis System (DFRAS) API
    
    A comprehensive AI-powered platform for analyzing delivery failures and providing predictive insights for logistics operations.
    
    ## Key Features
    - Multi-source data integration (orders, fleet logs, warehouse records, customer feedback)
    - AI-powered root cause analysis using LLM (all-MiniLM-L6-v2)
    - Automated correlation engine for pattern recognition
    - Intelligent reporting with actionable recommendations
    - Predictive analytics for proactive failure prevention
    
    ## Authentication
    All endpoints require JWT authentication. Use the `/auth/login` endpoint to obtain a token.
    
    ## Sample Data
    The system includes comprehensive sample data from `third-assignment-sample-data-set`:
    - 750+ clients across multiple cities
    - 15,000+ orders with delivery status
    - 50+ warehouses with capacity data
    - 200+ drivers with performance metrics
    - External factors (weather, traffic, events)
    - Customer feedback and ratings
  version: 1.0.0
  contact:
    name: DFRAS Support
    email: support@dfras.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local Development Server
  - url: http://localhost:8002
    description: Analytics Service Direct Access
  - url: http://localhost:8001
    description: Data Service Direct Access
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags:
        - Authentication
      summary: User Login
      description: Authenticate user and obtain JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - username
                - password
              properties:
                username:
                  type: string
                  example: "admin"
                password:
                  type: string
                  example: "admin123"
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                    example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
                  token_type:
                    type: string
                    example: "bearer"
                  user:
                    type: object
                    properties:
                      id:
                        type: string
                      username:
                        type: string
                      role:
                        type: string
        '401':
          description: Invalid credentials
...
  /api/data-ingestion/sample-data:
    post:
      tags:
        - Data Ingestion
      summary: Ingest Sample Data
      description: Populate database with sample data from third-assignment-sample-data-set
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties: {}
      responses:
        '200':
          description: Sample data ingested successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  message:
                    type: string
                    example: "Sample data ingested successfully"
                  records_processed:
                    type: integer
                    example: 15000
                  tables_populated:
                    type: array
                    items:
                      type: string
                    example: ["clients", "orders", "warehouses", "drivers"]
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...